add my part

+++ src/model/IGraph.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface IGraph<T> {
 	
 	public void add(T toAdd);
-	public void connect(T one, T two, double weight);
+	public void connect(T one, T two, int weight);
 	public void delete(T toDelete);
 	public T consult(String theVertex);
 	public void BFS(String origin);
